refactor: improve some components in companies profile and fix minor issues
