+++ apps/api/src/validations/company-profile-update.validation.ts
@@ -74,7 +74,10 @@ const companyProfileUpdateSchema = () => {
       .optional()
       .min(8, 'Telp must be at least 8 characters')
       .max(15, 'Telp must be at most 15 characters')
-      .matches(/^\d+$/, 'Nomor telepon hanya boleh berisi angka'),
+      .matches(
+        /^[\d\s()+-]+$/,
+        'Nomor telepon hanya boleh berisi angka, spasi, (), +, dan -',
+      ),
     description: Yup.string().optional(),
     category: Yup.string().optional(),
     location: Yup.string().optional(),
