+++ apps/web/src/app/(main)/page.tsx
@@ -7,7 +7,7 @@ import { NearestJobs } from './_components/nearest-job';
 import { useSearchParams } from 'next/navigation';
 import { useEffect, useState } from 'react';
 import AppLoading from '@/components/ui/app-loading';
-import { Job, JobsResponse } from '@/context/search-job-context';
+import { Job, JobsResponse, useSearchJob } from '@/context/search-job-context';
 import { axiosPublic } from '@/lib/axios';
 import { useToast } from '@/hooks/use-toast';
 
@@ -29,10 +29,12 @@ function deleteCookie(name: string) {
 }
 export default function HomePage() {
   const [loading, setLoading] = useState(true);
+  const { fetchJobs } = useSearchJob();
   const searchParams = useSearchParams();
   const urlMessage = searchParams.get('message');
   const [jobs, setJobs] = useState<Job[]>([]);
   const { toast } = useToast();
+  const [isFetch, setIsFetch] = useState(false);
   useEffect(() => {
     setLoading(true);
     const fetchJobs = async () => {
@@ -55,6 +57,36 @@ export default function HomePage() {
 
     fetchJobs();
   }, [toast]);
+
+  useEffect(() => {
+    const dateFilterParam = searchParams.get('dateFilter');
+    const sortOrderParam =
+      (searchParams.get('sortOrder') as 'asc' | 'desc') || 'desc';
+    const dateFromParams = searchParams.get('startDate');
+    const dateToParams = searchParams.get('endDate');
+    navigator.geolocation.getCurrentPosition(async (position) => {
+      try {
+        setIsFetch(true);
+        fetchJobs({
+          title: '',
+          category: '',
+          location: position.coords.latitude + ',' + position.coords.longitude,
+          dateFilter: dateFilterParam as any,
+          startDate: dateFromParams ? new Date(dateFromParams) : null,
+          endDate: dateToParams ? new Date(dateToParams) : null,
+          sortOrder: sortOrderParam,
+        });
+      } catch (error) {
+        toast({
+          title: 'Error',
+          description: 'Something went wrong',
+          variant: 'destructive',
+        });
+      } finally {
+        setIsFetch(false);
+      }
+    });
+  }, [fetchJobs, searchParams, toast]);
 
   useEffect(() => {
     const timer = setTimeout(() => {
@@ -80,14 +112,14 @@ export default function HomePage() {
   }, [urlMessage, toast]);
 
   return loading ? (
-    <div className="bg-background fixed left-0 top-0 flex min-h-screen w-screen flex-1 items-center justify-center">
+    <div className="fixed top-0 left-0 flex items-center justify-center flex-1 w-screen min-h-screen bg-background">
       <AppLoading size="md" label="Loading data, please stand by..." />
     </div>
   ) : (
     <Container className="">
-      <div className="flex w-full flex-col items-center justify-center">
+      <div className="flex flex-col items-center justify-center w-full">
         <div className="font-geist mb-[5px] flex h-fit w-full flex-col items-center justify-center">
-          <SearchBar />
+          <SearchBar isFetch={isFetch} />
         </div>
         <NewestJobs jobs={jobs} />
         <NearestJobs />
